+++ docs/CHANGELOG.rst
@@ -6,6 +6,7 @@ v0.1.33 (2025-10-20)
 ------------------------------------------
 
 * Update repo infrastructure
+* Update to newest MDAnalysis version
 * Update docs design
 
 v0.1.32 (2024-07-09)