+++ docs/CHANGELOG.rst
@@ -5,9 +5,9 @@ Changelog
 v0.1.33 (2025-10-20)
 ------------------------------------------
 
-* Update repo infrastructure
 * Update to newest MDAnalysis version
-* Update docs design
+* Update docs design and deployment
+* Update repo infrastructure
 
 v0.1.32 (2024-07-09)
 ------------------------------------------