+++ docs/CHANGELOG.rst
@@ -6,7 +6,7 @@ v0.1.33 (2025-10-20)
 ------------------------------------------
 
 * Update to newest MDAnalysis version
-* Update docs design and deployment
+* Update documentation design and deployment
 * Update repo infrastructure
 
 v0.1.32 (2024-07-09)